split the original `data.js` routes into individual express.js middlewares -- soho-7803
