+++ app/data/dummy-dropdown-data.js
@@ -1,0 +1,6 @@
+const data = require('../src/js/get-junk-dropdown-data');
+
+module.exports = (req, res, next) => {
+  res.json(data);
+  next();
+};
